Fix rowsAffected in IsUpToDate function
RowsAffected() does not return the right amount of rows affected.
Temporarily storing the results in an array of interface{} and then
checking the length of the temporary array yields the total number of
rows returned. If there is >= 1 row returned, there would be rows of
data that isn't updated accordingly and as such, must be updated