Fix bug where data is inaccurate
For certain countries where the names/states may be substrings of other
countries, as such, the time series will return the wrong value for the
record as it will find the first match which could be a country with the
substring. Thus, the search for the country/state must be strict to
ensure that this issue does not occur.